+++ projects/core/src/lib/storage/storage.service.ts
@@ -1,7 +1,7 @@
 import { inject, Injectable } from '@angular/core';
 
 import { PlatformService } from '../services';
-import { Nullable } from '../types';
+import { Nullable } from '@rt-tools/utils';
 import { STORAGE_TYPES_ENUM, StorageType } from './enums/storage-types.enum';
 import { IStorageConfig } from './interfaces/storage-config';
 import { JsonConverter } from './json-converter';