+++ projects/ui-kit/src/lib/ui-kit/table/components/table-container/table-container.component.ts
@@ -33,7 +33,8 @@ import { MatTooltip } from '@angular/material/tooltip';
 import { DomSanitizer } from '@angular/platform-browser';
 import { debounceTime, distinctUntilChanged, filter, map } from 'rxjs/operators';
 
-import { BlockDirective, ElemDirective, ModDirective, Nullable, PlatformService, WINDOW } from '@rt-tools/core';
+import { BlockDirective, ElemDirective, ModDirective, PlatformService, WINDOW } from '@rt-tools/core';
+import { Nullable } from '@rt-tools/utils';
 import { isString, PageModel, transformStringInput } from '@rt-tools/utils';
 import { BreakpointService, RtIconOutlinedDirective } from '@rt-tools/core';
 import { RtAsideService } from '../../../aside';